requirements.txt
# openai library
openai

# http client
requests

# simply ui framework
streamlit

# .env utility
python-dotenv

# jupyter notbook
jupyter

# finance
yfinance 
pandas 
numpy
matplotlib
mplfinance
pytz
nltk
pandas_datareader

# pdf
pyPDF2
pdfplumber

#extract
camelot-py[base]
pandas
tabulate
pyhwp
b64encode
plotly.graph_objects
plotly.express
sqlite
